+++ src/Symfony/Component/Translation/Resources/functions.php
@@ -14,7 +14,7 @@ namespace Symfony\Component\Translation;
 /**
  * @author Nate Wiebe <nate@northern.co>
  */
-function t(string $message, array $parameters = [], string $domain = null): Translatable
+function t(string $message, array $parameters = [], string $domain = null): TranslatableMessage
 {
-    return new Translatable($message, $parameters, $domain);
+    return new TranslatableMessage($message, $parameters, $domain);
 }