+++ src/Symfony/Component/Translation/Tests/fixtures/extractor/translatable.html.php
@@ -1,32 +1,32 @@
 This template is used for translation message extraction tests
-<?php new Translatable('translatable single-quoted key'); ?>
-<?php new Translatable('translatable double-quoted key'); ?>
-<?php new Translatable(<<<EOF
+<?php new TranslatableMessage('translatable single-quoted key'); ?>
+<?php new TranslatableMessage('translatable double-quoted key'); ?>
+<?php new TranslatableMessage(<<<EOF
 translatable heredoc key
 EOF
 ); ?>
-<?php new Translatable(<<<'EOF'
+<?php new TranslatableMessage(<<<'EOF'
 translatable nowdoc key
 EOF
 ); ?>
-<?php new Translatable(
+<?php new TranslatableMessage(
     "translatable double-quoted key with whitespace and escaped \$\n\" sequences"
 ); ?>
-<?php new Translatable(
+<?php new TranslatableMessage(
     'translatable single-quoted key with whitespace and nonescaped \$\n\' sequences'
 ); ?>
-<?php new Translatable(<<<EOF
+<?php new TranslatableMessage(<<<EOF
 translatable heredoc key with whitespace and escaped \$\n sequences
 EOF
 ); ?>
-<?php new Translatable(<<<'EOF'
+<?php new TranslatableMessage(<<<'EOF'
 translatable nowdoc key with whitespace and nonescaped \$\n sequences
 EOF
 ); ?>
 
-<?php new Translatable('translatable single-quoted key with "quote mark at the end"'); ?>
+<?php new TranslatableMessage('translatable single-quoted key with "quote mark at the end"'); ?>
 
-<?php new Translatable('translatable concatenated'.' message'.<<<EOF
+<?php new TranslatableMessage('translatable concatenated'.' message'.<<<EOF
  with heredoc
 EOF
 .<<<'EOF'
@@ -34,14 +34,14 @@ EOF
 EOF
 ); ?>
 
-<?php new Translatable('translatable other-domain-test-no-params-short-array', [], 'not_messages'); ?>
+<?php new TranslatableMessage('translatable other-domain-test-no-params-short-array', [], 'not_messages'); ?>
 
-<?php new Translatable('translatable other-domain-test-no-params-long-array', [], 'not_messages'); ?>
+<?php new TranslatableMessage('translatable other-domain-test-no-params-long-array', [], 'not_messages'); ?>
 
-<?php new Translatable('translatable other-domain-test-params-short-array', ['foo' => 'bar'], 'not_messages'); ?>
+<?php new TranslatableMessage('translatable other-domain-test-params-short-array', ['foo' => 'bar'], 'not_messages'); ?>
 
-<?php new Translatable('translatable other-domain-test-params-long-array', ['foo' => 'bar'], 'not_messages'); ?>
+<?php new TranslatableMessage('translatable other-domain-test-params-long-array', ['foo' => 'bar'], 'not_messages'); ?>
 
-<?php new Translatable('translatable typecast', ['a' => (int) '123'], 'not_messages'); ?>
+<?php new TranslatableMessage('translatable typecast', ['a' => (int) '123'], 'not_messages'); ?>
 
-<?php new Translatable('translatable default domain', [], null); ?>
+<?php new TranslatableMessage('translatable default domain', [], null); ?>